requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
distro==1.9.0
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.35.10
pocketsphinx==5.0.3
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
requests==2.32.3
setuptools==70.2.0
sniffio==1.3.1
sounddevice==0.4.7
SpeechRecognition==3.10.4
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.43.0
